Rental fixes

+++ src/modules/property/useCase/DeleteProperty/DeletePropertyUseCase.ts
@@ -1,4 +1,5 @@
 import { IPropertiesRepository } from "@modules/property/repositories/IPropertiesRepository"
+import { AppError } from "@shared/errors/AppError";
 import { inject, injectable } from "tsyringe"
 
 @injectable()
@@ -10,6 +11,12 @@ class DeletePropertyUseCase {
     ) {}
 
     async execute(id: string): Promise<void> {
+        const propertyExists = await this.propertiesRepository.findById(id);
+
+        if(!propertyExists) {
+            throw new AppError("Property doesn't exist!")
+        }
+
         await this.propertiesRepository.delete(id)
         return
     }
